move media folder, make path to media a constant
