+++ src/snakeGame/userInterface/Score.java
@@ -1,4 +1,4 @@
-package snakeGame;
+package snakeGame.userInterface;
 
 import javafx.scene.Group;
 import javafx.scene.control.Label;
