+++ src/snakeGame/userInterface/media/splash.mp4
